ft_transcendence/project/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels_redis
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.4
django-allauth==0.62.1
django-cors-headers
django-extensions
django-formtools==2.5.1
django-otp==1.5.0
django-phonenumber-field==7.3.0
django-two-factor-auth==1.16.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
oauthlib==3.2.2
phonenumbers==8.13.35
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
pypng==0.20220715.0
python-dotenv==1.0.1
python3-openid==3.2.0
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
daphne
Twisted[http2,tls]

